refactor all project
