+++ src/public/src/scripts/react/components/navbar/navbarAdmin.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import {Link} from "react-router";
-import sessionStore from "../../stores/sessionStore";
+import sessionStore from "../../../stores/sessionStore";
 
 
 /**
