+++ src/public/src/scripts/react/pages/common/error403.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
-import Navbar from "../navbar/navbar";
+import Navbar from "../../components/navbar/navbar";
 import {browserHistory} from "react-router";
-import {PermissionLevel} from "../../stores/sessionStore";
-import ErrorMessage from "./errorMessageComponent";
-import sessionActionCreator from "../../actions/sessionActionCreator";
+import {PermissionLevel} from "../../../stores/sessionStore";
+import ErrorMessage from "./../../components/errorMessageComponent";
+import sessionActionCreator from "../../../actions/sessionActionCreator";
 
 
 /**
