+++ src/public/src/scripts/react/pages/common/profile.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 import {Link, browserHistory} from "react-router";
 import * as ReactDOM from "react-dom";
-import ErrorMessage from "../errorMessageComponent";
-import Navbar from "../../navbar/navbar";
+import ErrorMessage from "../../components/errorMessageComponent";
+import Navbar from "../../components/navbar/navbar";
 import sessionStore, {PermissionLevel} from "../../../stores/sessionStore";
 import userStore from "../../../stores/userStore";
 import companyStore from "../../../stores/companyStore";
