+++ src/public/src/scripts/react/pages/sa/addCompany.tsx
@@ -1,95 +1,96 @@
 import * as React from "react";
-import {Link, hashHistory} from "react-router";
+import {browserHistory} from "react-router";
 import * as ReactDOM from "react-dom";
-import Navbar from "../../navbar/navbar";
+import Navbar from "../../components/navbar/navbar";
 import sessionStore, {PermissionLevel} from "../../../stores/sessionStore";
-import userStore from "../../../stores/userStore";
-import ErrorMessage from "../errorMessageComponent";
-import userActionCreators, {ISupeAdminCreation}
-    from "../../../actions/userActionCreator";
-/**
- * This interface represents the state of the InviteSuperAdmin page
- */
-interface ICreateSuperAdminState {
+import companyActionCreator, {IAddCompanyUser, ICompanyName}
+    from "../../../actions/companyActionCreator";
+import ErrorMessage from "../../components/errorMessageComponent";
+import companyStore from "../../../stores/companyStore";
 
-    message : string,
-    token : string
+
+/**
+ * This interface represents the state of the {AddCompany} page.
+ */
+export interface IAddCompanyState {
+    message : string;
+    token : string;
 }
 
 /**
- * <p>This class represents the super admin creation page.</p>
+ * <p>This class represents the add company page.</p>
  *
  * @history
  * | Author           | Action Performed               | Data       |
  * |------------------|--------------------------------|------------|
- * | Davide Polonio    | Create interfaces and class    | 15/06/2016 |
+ * | Davide Rigoni    | Create interfaces and class    | 06/06/2016 |
  *
- * @author  Davide Polonio
+ * @author  Davide Rigoni
  * @license MIT
  *
  */
-class AddSuperAdmin extends React.Component<void, ICreateSuperAdminState> {
-
+class AddCompany extends React.Component<void, IAddCompanyState> {
 
     /**
-     * @description Default constructor.
-     * @return {AddSuperAdmin}
+     * @description
+     * <p>This constructor calls his super constructor.
+     * It creates an AddCompany, defines its state and
+     * binds _onChange function to "this"</p>
+     * @return {AddCompany}
      */
     constructor() {
         super();
-        let superAdminErrorMessage : string = "";
-        if (userStore.isSuperAdminCreationErrored()) {
-            superAdminErrorMessage =
-              userStore.getSuperAdminCreationErrorMessage();
-        }
         this.state = {
-            message : superAdminErrorMessage,
-            token : sessionStore.getAccessToken()
+            message: companyStore.getAddCompanyError(),
+            token: sessionStore.getAccessToken()
         };
 
         this._onChange = this._onChange.bind(this);
-
     }
 
     /**
      * @description
      * <p>Render method of the component.
-     * It renders the AddSuperAdmin component.</p>
+     * It renders the AddCompany component.</p>
      * @return {JSX.Element}
      */
     public render() : JSX.Element {
-
         /* tslint:disable: max-line-length */
         return(
             <div>
                 <Navbar />
                 <div id="contentBody" className="container">
                     <div id="titles">
-                        <h3>Add a new Super Admin</h3>
+                        <h3>Add company</h3>
                     </div>
                     <div className="divider"></div>
-
                     <div className="row">
                         <ErrorMessage error={this.state.message} />
                         <form className="col s12">
                             <div className="row">
                                 <div className="input-field col s12">
+                                    <input id="companyName" type="text" className="validate" ref="companyName"/>
+                                    <label for="companyName">Name of the company</label>
+                                </div>
+                            </div>
+                            <div className="row">
+                                <div className="input-field col s12">
                                     <i className="material-icons prefix">email</i>
-                                    <input id="email" type="email" className="validate" ref="email"/>
-                                    <label for="email">Email</label>
+                                    <input id="email" type="email" className="validate"  ref="email"/>
+                                    <label for="email">Email of the Owner</label>
                                 </div>
                             </div>
                             <div className="row">
                                 <div className="input-field col s12">
                                     <i className="material-icons prefix">lock</i>
                                     <input id="password" type="text" className="validate"  ref="password"/>
-                                    <label for="password">Password of the new Super Admin</label>
+                                    <label for="password">Password of the Owner</label>
                                 </div>
                             </div>
                             <div className="right">
-                                <a className="waves-effect waves-light btn" onClick={this.addSuperAdmin.bind(this)}>
+                                <a className="waves-effect waves-light btn" onClick={this.addCompany.bind(this)}>
                                     <i className="material-icons left">done</i>
-                                    Create Super Admin
+                                    Add Company
                                 </a>
                             </div>
                         </form>
@@ -102,65 +103,58 @@ class AddSuperAdmin extends React.Component<void, ICreateSuperAdminState> {
 
     /**
      * @description
-     * <p>This method is call when the user click on the Add Super Admin
-     * button. A new action is created.</p>
+     * <p>This method is call when the user click on the add Company button.</p>
      */
-    private addSuperAdmin() : void {
-
-        console.log("On addSuperAdmin!");
+    private addCompany() : void {
         let email : string =
             ReactDOM.findDOMNode<HTMLInputElement>(this.refs["email"]).value;
-
         let password : string =
             ReactDOM.findDOMNode<HTMLInputElement>(this.refs["password"]).value;
-
-        let adminToCreate : ISupeAdminCreation = {
-
-            email : email,
-            password : password
-        };
-
-        console.log("E-mail: " + email);
-        console.log("Password: " + password);
-
-        // Creating a new action
-        userActionCreators.addSuperAdmin(adminToCreate, this.state.token);
-
+        let companyName : string =
+            ReactDOM.
+            findDOMNode<HTMLInputElement>(this.refs["companyName"]).value;
+        companyActionCreator.addCompany(
+            {
+                email : email,
+                password : password
+            },
+            {
+                name : companyName
+            }
+        );
     }
+
     /**
      * @description This method is called when the component mount.
      */
     private componentDidMount() : void {
         if (!(sessionStore.checkPermission(PermissionLevel.SUPERADMIN))) {
-            hashHistory.push("/Error403")
+            browserHistory.push("/Error403")
         }
-        userStore.addChangeListener(this._onChange)
+        companyStore.addChangeListener(this._onChange);
     }
 
     /**
      * @description This method is called when the component will unmount.
      */
     private componentWillUnmount() : void {
-        userStore.removeChangeListener(this._onChange);
+        companyStore.removeChangeListener(this._onChange);
     }
 
     /**
      * @description This method is called every time the store change.
      */
     private _onChange() : void {
-
         let errorMessage : string = "";
-
-        if ( userStore.isSuperAdminCreationErrored() ) {
-
-            errorMessage = userStore.getSuperAdminCreationErrorMessage();
+        if (companyStore.addCompanyError()) {
+            errorMessage = companyStore.getAddCompanyError()
         }
         this.setState({
-
             message : errorMessage,
             token : sessionStore.getAccessToken()
         });
     }
 }
 
-export default AddSuperAdmin;
+
+export default AddCompany;
