+++ src/public/src/scripts/react/pages/sa/showSuperAdmins.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import {Link, browserHistory} from "react-router";
-import Navbar from "../../navbar/navbar";
+import Navbar from "../../components/navbar/navbar";
 import sessionStore, {PermissionLevel} from "../../../stores/sessionStore";
 import userActionCreators,
         {ISuperAdmin} from "../../../actions/userActionCreator";
