+++ src/public/src/scripts/stores/companyStore.ts
@@ -431,7 +431,7 @@ class CompanyStore extends EventEmitter {
     }
 
     /**
-     * @description Emit changes to React components.
+     * @description Emit changes to React pages.
      * @returns {void}
      */
     private emitChange() : void {
