+++ src/public/src/scripts/stores/userStore.ts
@@ -416,7 +416,7 @@ class UserStore extends EventEmitter {
     }
 
     /**
-     * @description Emit changes to React components.
+     * @description Emit changes to React pages.
      * @returns {void}
      */
     private emitChange() : void {
